# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "operationId": "AdminUsers_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "`field:asc` or `field:desc`; allowed fields are listed per endpoint",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UserStatus"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches username, display name or email (case-insensitive substring). Sort fields: username, createdAt, updatedAt, role, status.",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPageDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "List users",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminUsers_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Username taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Create a user",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsers_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Get a user",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "description": "Disabling a user revokes their tokens. You cannot demote or disable yourself, nor the last active administrator (409).",
        "operationId": "AdminUsers_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "SELF_MODIFICATION or LAST_ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "description": "Deletes their personal address book and tokens. Shared books they created remain.",
        "operationId": "AdminUsers_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "SELF_MODIFICATION or LAST_ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/password": {
      "post": {
        "description": "Revokes every token of the user.",
        "operationId": "AdminUsers_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Reset a password",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/tokens": {
      "get": {
        "operationId": "AdminUsers_listTokens",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "`field:asc` or `field:desc`; allowed fields are listed per endpoint",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPageDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "List a user's tokens (RustDesk clients and admin sessions), newest first",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "RustDesk API server",
    "description": "Self-hosted replacement for the RustDesk **API server** (port 21114). It does not replace `hbbs`/`hbbr`.\n\n**rustdesk** — endpoints the RustDesk client calls. Quirks of the client contract:\n- bodies are parsed as JSON whatever the Content-Type; an empty body is `{}`;\n- errors are `{\"error\": \"<string>\"}`;\n- address-book mutations and audit posts answer **200 with an empty body** (204 or a non-empty body is a failure for the client);\n- `/api/sysinfo` and `/api/sysinfo_ver` answer plain text;\n- audit, heartbeat and sysinfo are unauthenticated by protocol.\n\n**admin** — the admin panel API under `/api/admin`: session cookie, role ADMIN, an allowed `Origin` on state-changing requests,\nerrors `{\"error\": {\"code\", \"message\", \"details\"?}}`, ISO-8601 UTC timestamps, durations in whole seconds.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token from POST /api/login"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rd_admin_session"
      }
    },
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "INVALID_CREDENTIALS",
          "TWO_FACTOR_UNSUPPORTED",
          "USER_DISABLED",
          "FORBIDDEN",
          "CSRF_ORIGIN_MISMATCH",
          "DEVICE_NOT_ALLOWED",
          "NOT_FOUND",
          "CONFLICT",
          "ALREADY_EXISTS",
          "LAST_ADMIN",
          "SELF_MODIFICATION",
          "AB_READ_ONLY",
          "AB_PEER_LIMIT",
          "SESSION_NOT_ACTIVE",
          "RANGE_TOO_LARGE",
          "PAYLOAD_TOO_LARGE",
          "RATE_LIMITED",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "username"
          },
          "message": {
            "type": "string",
            "example": "username must be longer than or equal to 2 characters"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "AdminErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "example": "NOT_FOUND",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "User not found"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "AdminErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/AdminErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "USER"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "DISABLED"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string",
            "description": "Lower-cased login name."
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "note",
          "role",
          "status",
          "createdAt",
          "updatedAt",
          "username"
        ]
      },
...
      "UserPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "total": {
            "type": "number",
            "example": 1
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "2-64 characters: letters, digits, `.`, `_`, `-`. Stored lower-cased.",
            "example": "jdoe"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256,
            "writeOnly": true
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 254,
            "format": "email"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "role": {
            "default": "USER",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "status": {
            "default": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 254,
            "format": "email"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          }
        }
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256,
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ]
      },
      "TokenKind": {
        "type": "string",
        "enum": [
          "RUSTDESK_CLIENT",
          "ADMIN_WEB"
        ]
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The token's `jti`.",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TokenKind"
              }
            ]
          },
          "clientId": {
            "type": "string",
            "nullable": true,
            "description": "RustDesk ID of the client that logged in"
          },
          "clientUuid": {
            "type": "string",
            "nullable": true,
            "description": "Machine UUID of the client that logged in"
          },
          "deviceInfo": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "revokedReason": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean",
            "description": "Neither revoked nor expired."
          },
          "current": {
            "type": "boolean",
            "description": "The token of the request that fetched this list (the caller's own admin session)."
          }
        },
        "required": [
          "id",
          "userId",
          "kind",
          "clientId",
          "clientUuid",
          "deviceInfo",
          "ip",
          "userAgent",
          "issuedAt",
          "expiresAt",
          "lastUsedAt",
          "revokedAt",
          "revokedReason",
          "active",
          "current"
        ]
      },
      "TokenPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenDto"
            }
          },
          "total": {
            "type": "number",
            "example": 1
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
...
    }
  }
}